pyproject.toml
[project]
name = "llm-factor-recsys-hnm"
version = "0.1.0"
description = "3-Layer Attribute Taxonomy + KAR Hybrid-Expert Adaptor for H&M Fashion Recommendation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Core ML (JAX ecosystem)
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "flax>=0.8.0",
    "optax>=0.1.7",
    "grain>=0.2.0",
    "orbax-checkpoint>=0.4.0",

    # Text Encoding
    "sentence-transformers>=2.2.0",
    "transformers>=4.36.0",

    # LLM API
    "openai>=1.6.0",

    # Vector Search
    "faiss-cpu>=1.7.4",

    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
    "duckdb>=0.9.0",

    # Traditional ML / Clustering / Baselines
    "scikit-learn>=1.3.0",
    "umap-learn>=0.5.0",
    "hdbscan>=0.8.33",
    "implicit>=0.7.0",
    "scipy>=1.11.0",

    # Experiment Tracking
    "wandb>=0.16.0",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",       # 인터랙티브 대시보드(Streamlit)만
    "streamlit>=1.29.0",

    # Config Management
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",

    # Parallel Processing
    "ray>=2.9.0",

    # Utilities
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.1.0",
]

notebook = [
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.1.0",
    "nbformat>=5.9.0",
]

serving = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
    "prometheus-client>=0.19.0",
    "prometheus-fastapi-instrumentator>=6.1.0",
]

pipeline = [
    "dvc>=3.30.0",
    "dvc-s3>=3.0.0",
]

loadtest = [
    "locust>=2.20.0",
]

all = [
    "llm-factor-recsys-hnm[dev,serving,pipeline,loadtest]",
]

gpu = [
    "jax[cuda12]>=0.4.20",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
